Physics Engine
Started on physics engine. Changed to Bullet, as it has been around
longer and is probably better. Got basic physics working, with 2 spheres
falling, one on top of the other, and them reacting to each other
(currently using same floor shader for spheres, as I couldn't work out
how to draw spheres using current setup, and glu draw function wasnt
working (can still tell that it is working as intended)).

Next to work on is to tidy code up, alter it so data is read from script
(or other source) and learn to draw different shapes for testing.
Down the track is to move current code out of main.cpp and main.h, so as
to maintain the Model-View-Controller design we need.

Note: Look right, and up to see spheres (floor planes?). Gravity is
currently set at -2, and that is changeable in Constructor of
PhysicsEngine.cpp.

Note 2: Don't forget doxygen comments!