Camera collision
Camera collision semi works. The camera movement is passed into the
physics engine, where a force is applied in the direction the camera was
moved. Calculations are done, then updated position is passed back for
drawing (moving) of camera location. All forces and velocities are wiped
at the start of each simulation step, so forces from previous movement
don't stay.

Known bugs:
Camera becomes stuck in place if it doesn't move for a short period of
time. Restart application to fix this.
If left long enough, second box stays in the air even after moving the
box below it out the way. Have an idea of what is causing this, but will
fix later.