Lua Script loading
Added in the script manager which i'll refine later tonight.
I've omitted the boost directory in the lua includes file as it seemed
to be 7000+ of unnecessary .h and .cpp files.

PrepareCube function was causing a fatal error so I've also commented
that out and replaced it with some script testing.