Heightfield Terrain Shape
Seem to successfully be loading in heightfield data and creating a
terrain shape with it, but with no terrain visible, it's hard to be
certain. Will have to wait until a terrain is able to be displayed.

Known bugs: Sometimes when building the program, it will throw an
exception about the size of the heightfield data. Other times it does
not and works. Just rerun until it works. Will fix later.