Major Changes
- Fixed shader issue where it was drawing as ortho instead of
persepective. This was because the camera position wasnt being passed to
the mesh vertexs correctly to be used in the shader code.
- Also incorrect ordering when storing attributes in VBO was a huge
issue that was missed and holy shit It was a coincidence that I stumbled
accross it! thak god for that.
- NewModel class is the new implementation of the old Model class which
reads in a batch of meshes using ASSIMP. This works correctly, just need
to start working on texture uv mapping.
- Best working project so far
Todo:
- Need to work on texture loading to be used with .obj files
- Keep an eye on the LUA issue that wouldnt run on barebones system
- Should be able to start implementing terrain very soon!!!
:100: